backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-genai
openai
python-dotenv
resend
pydantic
python-multipart
